requirements.txt
appdirs==1.4.3
asgiref==3.2.5
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.4
django-allauth==0.41.0
django-cors-headers==3.2.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
filelock==3.0.12
idna==2.9
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.4
pycparser==2.20
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.11
